requirements.txt
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiohttp-jinja2==1.6
aiosignal==1.3.1
altgraph==0.17.4
async-timeout==4.0.3
attrs==23.2.0
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
et-xmlfile==1.1.0
filehash==0.2.dev1
frozenlist==1.4.1
hidapi==0.14.0
idna==3.7
ifaddr==0.2.0
importlib-metadata==7.1.0
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
multidict==6.0.5
netifaces==0.11.0
openpyxl==3.1.2
packaging==24.0
passlib==1.7.4
pdfminer.six==20231228
pigpio==1.78
pillow==10.3.0
psutil==5.9.8
pycparser==2.22
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyMCP2221A==1.4.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
smbus==1.1.post2
tinydb==4.8.0
tomlkit==0.12.5
xlrd==2.0.1
yarl==1.9.4
zeroconf==0.132.2
zipp==3.19.0
